teste modelo

+++ cadproj/models.py
@@ -1,12 +1,12 @@
 from django.db import models
-from django.db.models.fields import CharField, IntegerField
+from django.db.models.fields import CharField, FloatField
 
 
 class Projeto(models.Model):
     nome_do_projeto = CharField(max_length=150, blank=False)
     data_de_inicio = CharField(max_length=10, blank=False)
     data_de_fim = CharField(max_length=10, blank=False)
-    valor_projeto = IntegerField()
+    valor_projeto = FloatField()
     RISCO = (
         (0, 'Baixo'),
         (1, 'Medio'),
